egg hatching reworks

+++ include/egg_hatch.h
@@ -1,10 +1,12 @@
 #ifndef GUARD_EGG_HATCH_H
 #define GUARD_EGG_HATCH_H
 
+#include "constants/egg_hatch.h"
+
 void ScriptHatchMon(void);
 bool8 CheckDaycareMonReceivedMail(void);
 void EggHatch(void);
-u8 GetEggStepMultiplier(void);
+u32 GetEggStepMultiplier(void);
 u16 CountPartyAliveNonEggMons(void);
 
 #endif // GUARD_EGG_HATCH_H
